Move MSTS format support to separate assembly
