+++ OpenBVE.sln
@@ -42,6 +42,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CarXMLConvertor", "source\C
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Texture.Dds", "source\Plugins\Texture.Dds\Texture.Dds.csproj", "{EEC75E83-E284-44E0-B321-3F5C6512074F}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Formats.Msts", "source\Plugins\Formats.MsTs\Formats.Msts.csproj", "{E81B7BD8-A326-47D3-B7EE-E9C7D4D119FA}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -220,6 +222,16 @@ Global
 		{EEC75E83-E284-44E0-B321-3F5C6512074F}.Release|Mixed Platforms.Build.0 = Release|x86
 		{EEC75E83-E284-44E0-B321-3F5C6512074F}.Release|x86.ActiveCfg = Release|x86
 		{EEC75E83-E284-44E0-B321-3F5C6512074F}.Release|x86.Build.0 = Release|x86
+		{E81B7BD8-A326-47D3-B7EE-E9C7D4D119FA}.Debug|Any CPU.ActiveCfg = Debug|x86
+		{E81B7BD8-A326-47D3-B7EE-E9C7D4D119FA}.Debug|Mixed Platforms.ActiveCfg = Debug|x86
+		{E81B7BD8-A326-47D3-B7EE-E9C7D4D119FA}.Debug|Mixed Platforms.Build.0 = Debug|x86
+		{E81B7BD8-A326-47D3-B7EE-E9C7D4D119FA}.Debug|x86.ActiveCfg = Debug|x86
+		{E81B7BD8-A326-47D3-B7EE-E9C7D4D119FA}.Debug|x86.Build.0 = Debug|x86
+		{E81B7BD8-A326-47D3-B7EE-E9C7D4D119FA}.Release|Any CPU.ActiveCfg = Release|x86
+		{E81B7BD8-A326-47D3-B7EE-E9C7D4D119FA}.Release|Mixed Platforms.ActiveCfg = Release|x86
+		{E81B7BD8-A326-47D3-B7EE-E9C7D4D119FA}.Release|Mixed Platforms.Build.0 = Release|x86
+		{E81B7BD8-A326-47D3-B7EE-E9C7D4D119FA}.Release|x86.ActiveCfg = Release|x86
+		{E81B7BD8-A326-47D3-B7EE-E9C7D4D119FA}.Release|x86.Build.0 = Release|x86
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -232,6 +244,7 @@ Global
 		{06D89847-9C7A-47D5-8C7A-95AEBFFF5F1E} = {16553295-E70F-4596-AA78-848EEA576C4A}
 		{E9B64673-65B7-4D77-A4DB-4B441A7C9197} = {3D6660F8-52D2-4A30-8F1E-1922DDCF10F9}
 		{EEC75E83-E284-44E0-B321-3F5C6512074F} = {16553295-E70F-4596-AA78-848EEA576C4A}
+		{E81B7BD8-A326-47D3-B7EE-E9C7D4D119FA} = {16553295-E70F-4596-AA78-848EEA576C4A}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {50B52A67-69B1-4B19-A59C-FF058FD9EBDF}
