+++ source/ObjectViewer/ObjectViewer.csproj
@@ -84,8 +84,6 @@
     <Compile Include="Parsers\Ls3DGrpParser.cs" />
     <Compile Include="Parsers\Ls3DObjectParser.cs" />
     <Compile Include="Parsers\MSTS\ShapeParser.cs" />
-    <Compile Include="Parsers\MSTS\FileReaders.cs" />
-    <Compile Include="Parsers\MSTS\Tokens.cs" />
     <Compile Include="Parsers\WavefrontObjParser.cs" />
     <Compile Include="Plugins.cs" />
     <Compile Include="Renderer.Strings.cs" />
@@ -127,6 +125,10 @@
     <ProjectReference Include="..\OpenBveApi\OpenBveApi.csproj">
       <Project>{27134980-4415-4375-A564-40A9014DFA5F}</Project>
       <Name>OpenBveApi</Name>
+    </ProjectReference>
+    <ProjectReference Include="..\Plugins\Formats.MsTs\Formats.Msts.csproj">
+      <Project>{E81B7BD8-A326-47D3-B7EE-E9C7D4D119FA}</Project>
+      <Name>OpenBve.Formats.Msts</Name>
     </ProjectReference>
   </ItemGroup>
   <ItemGroup />
