--- source/OpenBVE/Parsers/Object/MSTS/Tokens.cs
@@ -1,144 +1,0 @@
-﻿namespace OpenBve
-{
-    /// <summary>The tokens from the Kuju compressed binary format</summary>
-    public enum KujuTokenID : uint
-    {
-        /*
-		* utils\FFEDIT\coreids.tok
-        * NOTE: This only includes object based tokens, not the extended list
-        * 
-		*/
-        error = 0,
-        comment = 1,
-        point,
-        vector,
-        quat,
-        normals,
-        normal_idxs,
-        points,
-        uv_point,
-        uv_points,
-        colour,
-        colours,
-        packed_colour,
-        image,
-        images,
-        texture,
-        textures,
-        light_material,
-        light_materials,
-        linear_key,
-        tcb_key,
-        linear_pos,
-        tcb_pos,
-        slerp_rot,
-        tcb_rot,
-        controllers,
-        anim_node,
-        anim_nodes,
-        animation,
-        animations,
-        anim,
-        lod_controls,
-        lod_control,
-        distance_levels_header,
-        distance_level_header,
-        dlevel_selection,
-        distance_levels,
-        distance_level,
-        sub_objects,
-        sub_object,
-        sub_object_header,
-        geometry_info,
-        geometry_nodes,
-        geometry_node,
-        geometry_node_map,
-        cullable_prims,
-        vtx_state,
-        vtx_states,
-        vertex,
-        vertex_uvs,
-        vertices,
-        vertex_set,
-        vertex_sets,
-        primitives,
-        prim_state,
-        prim_states,
-        prim_state_idx,
-        indexed_point_list,
-        point_list,
-        indexed_line_list,
-        indexed_trilist,
-        tex_idxs,
-        tri,
-        vertex_idxs,
-        flags,
-        matrix,
-        matrices,
-        hierarchy,
-        volumes,
-        vol_sphere,
-        shape_header,
-        shape,
-        shader_names,
-        shader_name,
-        texture_filter_names,
-        texture_filter_name,
-        sort_vectors,
-        uvop_arg_sets,
-        uvop_arg_set,
-        light_model_cfgs,
-        light_model_cfg,
-        uv_ops,
-        uvop_copy,
-        uv_op_share,
-        uv_op_copy,
-        uv_op_uniformscale,
-        uv_op_user_uninformscale,
-        uv_op_nonuniformscale,
-        uv_op_user_nonuninformscale,
-        uv_op_transform,
-        uv_op_user_transform,
-        uv_op_reflectxy,
-        uv_op_reflectmap,
-        uv_op_reflectmapfull,
-        uv_op_spheremap,
-        uv_op_spheremapfull,
-        uv_op_specularmap,
-        uv_op_embossbump,
-        user_uv_args,
-        io_dev,
-        io_map,
-        sguid,
-        dlev_cfg_table,
-        dlev_cfg,
-        subobject_shaders,
-        subobject_light_cfgs,
-        shape_named_data,
-        shape_named_data_header,
-        shape_named_geometry,
-        shape_geom_ref,
-        material_palette,
-        blend_config,
-        blend_config_header,
-        filtermode_cfgs,
-        filter_mode_cfg,
-        blend_mode_cfgs,
-        blend_mode_cfg,
-        texture_stage_progs,
-        texture_stage_prog,
-        blend_mode_cfg_refs,
-        shader_cfgs,
-        shader_cfg,
-        texture_slots,
-        texture_slot,
-        named_filter_modes,
-        named_filter_mode,
-        filtermode_cfg_refs,
-        filtermode_cfg_ref,
-        named_shaders,
-        named_shader,
-        shader_cfg_refs,
-        shader_cfg_ref,
-    }
-}
