+++ source/Plugins/Formats.Msts/Properties/AssemblyInfo.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Reflection;
+using System.Runtime.InteropServices;
+
+[assembly: AssemblyTitle("MSTS File Format Support Plugin")]
+[assembly: AssemblyDescription("Provides support functions for loading MSTS file formats")]
+[assembly: AssemblyProduct("openBVE")]
+[assembly: AssemblyCopyright("The openBVE Project")]
+[assembly: ComVisible(false)]
+[assembly: CLSCompliant(false)] //MSTS uses unsigned integers in the binary format, hence not CLS compliant
+
+[assembly: AssemblyVersion("0.0.0.0")]
