+++ source/Plugins/Formats.Msts/Tokens.cs
@@ -1,6 +1,6 @@
-﻿namespace OpenBve
+﻿namespace OpenBve.Formats.MsTs
 {
-    /// <summary>The tokens from the Kuju compressed binary format</summary>
+	/// <summary>The tokens from the Kuju compressed binary format</summary>
     public enum KujuTokenID : uint
     {
         /*
